# contracts/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Kanban API
  version: 1.1.0
...
paths:
...
  /api/boards:
    get:
      operationId: Boards_list
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 1
          explode: false
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 20
          explode: false
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                  - totalCount
                  - page
                  - pageSize
                  - totalPages
                  - hasNextPage
                  - hasPreviousPage
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Board'
                  totalCount:
                    type: integer
                    format: int32
                    description: Total number of records matching the query
                  page:
                    type: integer
                    format: int32
                    description: Current page number (1-based)
                  pageSize:
                    type: integer
                    format: int32
                    description: Number of items per page
                  totalPages:
                    type: integer
                    format: int32
                    description: Total number of pages
                  hasNextPage:
                    type: boolean
                    description: Whether there is a next page
                  hasPreviousPage:
                    type: boolean
                    description: Whether there is a previous page
                description: Generic paginated response wrapper
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Boards
    post:
      operationId: Boards_create
      parameters: []
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Board'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Boards
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBoardRequest'
  /api/boards/{boardId}:
    get:
      operationId: Boards_get
      parameters:
        - name: boardId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardDetail'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Boards
    patch:
      operationId: Boards_update
      parameters:
        - name: boardId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Board'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Boards
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBoardRequest'
    delete:
      operationId: Boards_delete
      parameters:
        - name: boardId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Boards
  /api/boards/{boardId}/columns:
    post:
      operationId: Columns_create
      parameters:
        - name: boardId
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Column'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Columns
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateColumnRequest'
  /api/boards/{boardId}/columns/{columnId}:
    patch:
      operationId: Columns_update
      parameters:
        - name: boardId
          in: path
          required: true
          schema:
            type: string
        - name: columnId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Column'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Columns
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateColumnRequest'
    delete:
      operationId: Columns_delete
      parameters:
        - name: boardId
          in: path
          required: true
          schema:
            type: string
        - name: columnId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Columns
...
security:
  - {}
  - ApiKeyAuth: []
components:
  schemas:
...
    Board:
      type: object
      required:
        - id
        - title
        - ownerId
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        ownerId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    BoardDetail:
      type: object
      required:
        - id
        - title
        - ownerId
        - createdAt
        - updatedAt
        - columns
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        ownerId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        columns:
          type: array
          items:
            $ref: '#/components/schemas/Column'
    Column:
      type: object
      required:
        - id
        - title
        - position
        - boardId
        - tasks
      properties:
        id:
          type: string
        title:
          type: string
        position:
          type: integer
          format: int32
        boardId:
          type: string
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/TaskItem'
    CreateBoardRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
        description:
          type: string
    CreateColumnRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
...
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
      description: Standard error response returned on failures
...
    TaskItem:
      type: object
      required:
        - id
        - title
        - position
        - columnId
        - priority
        - tags
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        position:
          type: integer
          format: int32
        columnId:
          type: string
        assigneeId:
          type: string
        priority:
          type: string
        tags:
          type: array
          items:
            type: string
        dueDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UpdateBoardRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
    UpdateColumnRequest:
      type: object
      properties:
        title:
          type: string
        position:
          type: integer
          format: int32
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: cookie
      name: session
servers:
  - url: http://localhost:5010
    description: Development server
    variables: {}